Land B2 ingest layer; snapshot in-progress B3/B4/B5 work
B2 delivers the Connections.csv parser, target registry, deterministic
company-match ladder, person resolution and review queue. The parser sniffs
for the header row and ships a standing negative control asserting that a
fixed skiprows=3 would have corrupted three of its four preamble fixtures.
Dates parse through an explicit month table rather than strptime('%b'),
which is locale-sensitive. R5's 100-case company fixture passes 100/100.

Renames the hedge fund entity from citadel_llc to citadel in targets.yaml
to match the frozen Firm enum, and makes the registry loader reject any
entity id that is not a Firm member. Substitutes core.ids.fold_accents for
R5's unidecode step per ruling C2, with the fixture still at 100/100.

787 tests passing. The one remaining failure is in the enrichment layer,
which is still being written.

+++ tests/test_graph_missing.py
@@ -93,7 +93,7 @@ def _drop_target_columns(targets, fraction, seed=11):
     """Mark a fraction of targets unharvested — the tri-state way to "drop"."""
     rng = np.random.default_rng(seed)
     ordered = sorted(targets, key=lambda t: t.target_id)
-    n_drop = int(round(fraction * len(ordered)))
+    n_drop = round(fraction * len(ordered))
     dropped = set(rng.choice(len(ordered), size=n_drop, replace=False).tolist())
     return [
         Target(target_id=t.target_id, firm=t.firm, harvested=(i not in dropped))
